+++ test/mocha/12-profile-agents-get-by-token.js
@@ -3,19 +3,30 @@
  */
 'use strict';
 
+const {CapabilityAgent} = require('@digitalbazaar/webkms-client');
 const {profileAgents} = require('bedrock-profile');
 const helpers = require('./helpers');
-const {config, util: {uuid}} = require('bedrock');
+const {util: {uuid}} = require('bedrock');
 const mockData = require('./mock.data');
 
-const kmsBaseUrl = `${config.server.baseUri}/kms`;
-
 describe('profileAgents getByToken API', () => {
+  // top-level application capability agent for creating meters
+  let capabilityAgent;
+  let kmsMeterCapability;
   // mock session authentication for delegations endpoint
   let passportStub;
   before(async () => {
     await helpers.prepareDatabase(mockData);
     passportStub = await helpers.stubPassport();
+
+    // top-level applications must create meters to associate with the
+    // creation of profile agents; the tests here reuse the same meter but
+    // applications can create as many as needed
+    const secret = 'b07e6b31-d910-438e-9a5f-08d945a5f676';
+    const handle = 'app';
+    capabilityAgent = await CapabilityAgent.fromSecret({secret, handle});
+    const {meterCapability} = await helpers.createMeter({capabilityAgent});
+    kmsMeterCapability = meterCapability;
   });
   after(() => {
     passportStub.restore();
@@ -26,9 +37,8 @@ describe('profileAgents getByToken API', () => {
     let error;
     let profileAgent;
     try {
-      ({profileAgent} = await profileAgents.create({
-        kmsBaseUrl, profileId, token
-      }));
+      ({profileAgent} = await profileAgents.create(
+        {profileId, token, kmsMeterCapability}));
     } catch(e) {
       error = e;
     }